change User PricePerPack field to type  float

+++ CranberryAPI/Models/ApplicationUser.cs
@@ -8,7 +8,7 @@ public class ApplicationUser : IdentityUser
 
   public string QuitDate { get; set; }
   public int AvgSmokedDaily { get; set; }
-  public int PricePerPack { get; set; }
+  public float PricePerPack { get; set; }
   public int CigsPerPack { get; set; }
 
   [JsonIgnore]
