+++ CranberryAPI/Models/UpdateUserDto.cs
@@ -5,6 +5,6 @@ public class UpdateUserDto
 {
   public string QuitDate { get; set; }
   public int AvgSmokedDaily { get; set; }
-  public int PricePerPack { get; set; }
+  public float PricePerPack { get; set; }
   public int CigsPerPack { get; set; }
 }
