+++ CranberryAPI/Models/UserProfileDto.cs
@@ -7,7 +7,7 @@ public class UserProfileDto
   public string UserName { get; set; }
   public string QuitDate { get; set; }
   public int AvgSmokedDaily { get; set; }
-  public int PricePerPack { get; set; }
+  public float PricePerPack { get; set; }
   public int CigsPerPack { get; set; }
 
   // public List<Journal> Journals { get; set; }
